@@ -1,71 +1,70 @@
 {
-    "name": "admin-panel-front",
-    "version": "0.0.1",
-    "license": "MIT",
+    "name": "theme-adminator",
+    "version": "0.0.0",
     "scripts": {
         "ng": "ng",
         "start": "ng serve",
-        "build": "ng build --prod",
+        "build": "ng build",
         "test": "ng test",
         "lint": "ng lint",
         "e2e": "ng e2e"
     },
     "private": true,
     "dependencies": {
-        "@amcharts/amcharts4": "^4.3.9",
-        "@amcharts/amcharts4-geodata": "^4.1.4",
-        "@angular/animations": "^5.1.0",
-        "@angular/common": "^5.1.0",
-        "@angular/compiler": "^5.1.0",
-        "@angular/core": "^5.1.0",
-        "@angular/forms": "^5.1.0",
-        "@angular/http": "^5.1.0",
-        "@angular/platform-browser": "^5.1.0",
-        "@angular/platform-browser-dynamic": "^5.1.0",
-        "@angular/router": "^5.1.0",
-        "@types/jquery": "^3.3.29",
-        "bootstrap": "^4.0.0",
-        "bootstrap-datepicker": "^1.7.1",
+        "@angular/animations": "~7.2.0",
+        "@angular/common": "~7.2.0",
+        "@angular/compiler": "~7.2.0",
+        "@angular/core": "~7.2.0",
+        "@angular/forms": "~7.2.0",
+        "@angular/platform-browser": "~7.2.0",
+        "@angular/platform-browser-dynamic": "~7.2.0",
+        "@angular/router": "~7.2.0",
+        "@fortawesome/fontawesome-free": "^5.8.1",
+        "@fullcalendar/core": "^4.0.2",
+        "@fullcalendar/daygrid": "^4.0.1",
+        "@fullcalendar/list": "^4.0.1",
+        "@fullcalendar/timegrid": "^4.0.1",
+        "bootstrap": "^4.3.1",
+        "bootstrap-datepicker": "^1.8.0",
         "chart.js": "^2.8.0",
-        "core-js": "^2.4.1",
-        "datatables": "^1.10.13",
+        "core-js": "^2.5.4",
+        "datatables.net": "^1.10.19",
+        "datatables.net-dt": "^1.10.19",
         "dateformat": "^3.0.3",
         "easy-pie-chart": "^2.1.7",
-        "fullcalendar": "^3.6.2",
         "http-build-query": "^0.7.0",
-        "jquery": "^3.3.1",
+        "jquery": "^3.4.0",
         "jquery-price-format": "git+https://github.com/flaviosilveira/Jquery-Price-Format.git",
         "jquery-sparkline": "^2.4.0",
-        "lodash": "^4.17.5",
-        "masonry-layout": "^4.2.0",
-        "moment": "^2.19.1",
+        "masonry-layout": "^4.2.2",
+        "moment": "^2.24.0",
         "perfect-scrollbar": "^1.4.0",
-        "popper.js": "^1.14.7",
-        "rxjs": "^5.5.6",
-        "themify-icons": "^1.0.0",
-        "tooltip.js": "^1.3.1",
-        "zone.js": "^0.8.19"
+        "popper.js": "^1.15.0",
+        "rxjs": "~6.3.3",
+        "tooltip-js": "^3.0.0",
+        "tslib": "^1.9.0",
+        "zone.js": "~0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "^1.5.6",
-        "@angular/compiler-cli": "^5.1.0",
-        "@angular/language-service": "^5.1.0",
-        "@fortawesome/fontawesome-free": "^5.8.1",
-        "@types/jasmine": "~2.8.3",
-        "@types/jasminewd2": "~2.0.2",
-        "@types/node": "~6.0.60",
-        "codelyzer": "^4.0.1",
-        "jasmine-core": "~2.8.0",
+        "@angular-devkit/build-angular": "~0.13.0",
+        "@angular/cli": "~7.3.8",
+        "@angular/compiler-cli": "~7.2.0",
+        "@angular/language-service": "~7.2.0",
+        "@types/jasmine": "~2.8.8",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/jquery": "^3.3.29",
+        "@types/node": "~8.9.4",
+        "codelyzer": "~4.5.0",
+        "jasmine-core": "~2.99.1",
         "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~2.0.0",
+        "karma": "~4.0.0",
         "karma-chrome-launcher": "~2.2.0",
-        "karma-cli": "~1.0.1",
-        "karma-coverage-istanbul-reporter": "^1.2.1",
-        "karma-jasmine": "~1.1.0",
+        "karma-coverage-istanbul-reporter": "~2.0.1",
+        "karma-jasmine": "~1.1.2",
         "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "~5.1.2",
-        "ts-node": "~3.2.0",
-        "tslint": "~5.9.1",
-        "typescript": "~2.5.3"
+        "protractor": "~5.4.0",
+        "ts-node": "~7.0.0",
+        "tslint": "~5.11.0",
+        "typescript": "~3.2.2"
     }
 }
