@@ -1,17 +1,16 @@
 import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
 import {RouterModule, Routes} from '@angular/router';
-import {LayoutCleanComponent} from './core/layouts/layout-clean/layout-clean.component';
-import {RedirectIfAuthenticatedGuard} from './shared/guards/redirect-if-authenticated/redirect-if-authenticated.guard';
+import {GoogleMapsComponent} from './modules/google-maps/google-maps.component';
+import {FormsComponent} from './modules/forms/forms.component';
+import {UiElementsComponent} from './modules/ui-elements/ui-elements.component';
+import {PageBlankComponent} from './modules/page-blank/page-blank.component';
 import {LayoutDefaultComponent} from './core/layouts/layout-default/layout-default.component';
 import {RedirectIfNotAuthenticatedGuard} from './shared/guards/redirect-if-not-authenticated/redirect-if-not-authenticated.guard';
-import {PageBlankComponent} from './modules/page-blank/page-blank.component';
-import {UiElementsComponent} from './modules/ui-elements/ui-elements.component';
-import {FormsComponent} from './modules/forms/forms.component';
-import {GoogleMapsComponent} from './modules/google-maps/google-maps.component';
+import {RedirectIfAuthenticatedGuard} from './shared/guards/redirect-if-authenticated/redirect-if-authenticated.guard';
+import {LayoutCleanComponent} from './core/layouts/layout-clean/layout-clean.component';
 
 const routes: Routes = [
-    { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
+    {path: '', redirectTo: '/dashboard', pathMatch: 'full'},
     {
         path: '', data: {base: true}, component: LayoutCleanComponent, canActivate: [RedirectIfAuthenticatedGuard], children: [
             {path: '', loadChildren: './modules/auth/auth.module#AuthModule'},
@@ -36,13 +35,8 @@ const routes: Routes = [
 ];
 
 @NgModule({
-    imports: [
-        CommonModule,
-        RouterModule.forRoot(routes)
-    ],
-    exports: [
-        RouterModule,
-    ],
+    imports: [RouterModule.forRoot(routes)],
+    exports: [RouterModule]
 })
 export class AppRoutingModule {
 }
