@@ -1,28 +1,35 @@
 import {async, TestBed} from '@angular/core/testing';
+import {RouterTestingModule} from '@angular/router/testing';
 import {AppComponent} from './app.component';
 
 describe('AppComponent', () => {
     beforeEach(async(() => {
         TestBed.configureTestingModule({
+            imports: [
+                RouterTestingModule
+            ],
             declarations: [
                 AppComponent
             ],
-        }).compileComponents().then();
+        }).compileComponents();
     }));
-    it('should create the app', async(() => {
+
+    it('should create the app', () => {
         const fixture = TestBed.createComponent(AppComponent);
         const app = fixture.debugElement.componentInstance;
         expect(app).toBeTruthy();
-    }));
-    it(`should have as title 'app'`, async(() => {
+    });
+
+    it(`should have as title 'theme-adminator'`, () => {
         const fixture = TestBed.createComponent(AppComponent);
         const app = fixture.debugElement.componentInstance;
-        expect(app.title).toEqual('app');
-    }));
-    it('should render title in a h1 tag', async(() => {
+        expect(app.title).toEqual('theme-adminator');
+    });
+
+    it('should render title in a h1 tag', () => {
         const fixture = TestBed.createComponent(AppComponent);
         fixture.detectChanges();
         const compiled = fixture.debugElement.nativeElement;
-        expect(compiled.querySelector('h1').textContent).toContain('Welcome to app!');
-    }));
+        expect(compiled.querySelector('h1').textContent).toContain('Welcome to theme-adminator!');
+    });
 });
