@@ -1,6 +1,9 @@
 import {Component, Input, OnInit} from '@angular/core';
 import * as moment from 'moment';
-import {Calendar} from 'fullcalendar';
+import { Calendar } from '@fullcalendar/core';
+import dayGridPlugin from '@fullcalendar/daygrid';
+import timeGridPlugin from '@fullcalendar/timegrid';
+import listPlugin from '@fullcalendar/list';
 
 declare let $: any;
 
@@ -9,44 +12,102 @@ declare let $: any;
     templateUrl: './full-calendar.component.html'
 })
 export class FullCalendarComponent implements OnInit {
-    _events: Array<any> = [];
-
     calendar: Calendar;
 
+    calendarEvents: Array<any> = [];
+
     constructor() {
+    }
+
+    @Input()
+    set events(events: Array<any>) {
+        this.calendarEvents = events;
+
+        if (this.calendarEvents) {
+            this.initCalendar();
+        }
     }
 
     ngOnInit() {
         this.initCalendar();
     }
 
-    @Input()
-    set events(events: Array<any>) {
-        this._events = events;
-
-        if (this._events) {
-            this.initCalendar();
-        }
-    }
-
     initCalendar() {
-        let events = this.handlerEvents();
+        // const events = this.handlerEvents();
+
+        let events = [
+            {
+                title: 'All Day Event',
+                start: '2018-01-01',
+            },
+            {
+                title: 'Long Event',
+                start: '2018-01-07',
+                end: '2018-01-10'
+            },
+            {
+                id: 999,
+                title: 'Repeating Event',
+                start: '2018-01-09T16:00:00'
+            },
+            {
+                id: 999,
+                title: 'Repeating Event',
+                start: '2018-01-16T16:00:00'
+            },
+            {
+                title: 'Conference',
+                start: '2018-01-11',
+                end: '2018-01-13'
+            },
+            {
+                title: 'Meeting',
+                start: '2018-01-12T10:30:00',
+                end: '2018-01-12T12:30:00'
+            },
+            {
+                title: 'Lunch',
+                start: '2018-01-12T12:00:00'
+            },
+            {
+                title: 'Meeting',
+                start: '2018-01-12T14:30:00'
+            },
+            {
+                title: 'Happy Hour',
+                start: '2018-01-12T17:30:00'
+            },
+            {
+                title: 'Dinner',
+                start: '2018-01-12T20:00:00'
+            },
+            {
+                title: 'Birthday Party',
+                start: '2018-01-13T07:00:00'
+            },
+            {
+                title: 'Click for Google',
+                url: 'http://google.com/',
+                start: '2018-01-28'
+            }
+        ];
 
         if (this.calendar) {
             this.calendar.destroy();
         }
 
-        this.calendar = new Calendar($('#full-calendar'), {
-            events: events,
-            timezone: 'UTC', // America/Sao_Paulo
+        this.calendar = new Calendar(document.getElementById('full-calendar'), {
+            plugins: [ dayGridPlugin, timeGridPlugin, listPlugin ],
+            events,
+            // timezone: 'UTC', // America/Sao_Paulo
             locale: 'en', // pt-br
             height: 800,
             editable: true,
-            // navLinks: true, // can click day/week names to navigate views
+            navLinks: true, // can click day/week names to navigate views
             selectable: true,
             eventLimit: false, // allow "more" link when too many events
             header: {
-                left: 'month,agendaWeek,agendaDay',
+                left: 'dayGridMonth,timeGridWeek,timeGridDay,listWeek',
                 center: 'title',
                 right: 'today prev,next',
             },
@@ -58,7 +119,7 @@ export class FullCalendarComponent implements OnInit {
     private handlerEvents() {
         moment.locale('pt-br');
 
-        return this._events.map((event: any) => {
+        return this.calendarEvents.map((event: any) => {
             if (event.start) {
                 event.start = moment(event.start);
             }
