@@ -1,8 +1,9 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 
 @Injectable()
 export class ConversationService {
 
-  constructor() { }
+    constructor() {
+    }
 
 }
