@@ -4,8 +4,8 @@ import {routing} from './email-routing.module';
 import {InboxComponent} from './inbox/inbox.component';
 import {FormsModule} from '@angular/forms';
 import {SharedModule} from '../../shared/shared.module';
-import { ComposeComponent } from './compose/compose.component';
-import { NavComponent } from './components/nav/nav.component';
+import {ComposeComponent} from './compose/compose.component';
+import {NavComponent} from './components/nav/nav.component';
 
 @NgModule({
     imports: [
