@@ -1,5 +1,5 @@
 export const environment = {
-    production: true,
-    apiUrl: 'http://localhost:80',
-    token: '',
+  production: true,
+  apiUrl: 'http://localhost:80',
+  token: '',
 };
