@@ -5,8 +5,8 @@ import {AppModule} from './app/app.module';
 import {environment} from './environments/environment';
 
 if (environment.production) {
-    enableProdMode();
+  enableProdMode();
 }
 
 platformBrowserDynamic().bootstrapModule(AppModule)
-    .catch(err => console.log(err));
+  .catch(err => console.error(err));
